Update CertificateRequest.php
A small enhancement that allows providing CSR in PEM format. Useful if CSR is generated on another system (like HSM).  This patch is required by finalizeOrder() function.

The only issue here is that $keyPair is not needed however the exception is triggered if not provided. My personal workaround is to set it to the same variable used for SecureHttpClient() which is the account key. I know, it is a dirty hack but works, and minor changes were needed to support PEM CSR with acmephp :)

This is a quick fix, more rework is needed. 
CommonName for $distinguishedName we can get by parsing PEM CSR with openssl_csr_get_subject();